feat: Atualizar limite de nome de tarefa
